% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kread-package.R
\docType{package}
\name{kread-package}
\alias{kread}
\alias{kread-package}
\title{kread: kread - R package for reading files}
\description{
The kread package provides functions to read various file formats into R, making data import easier and more efficient.
}
\author{
\strong{Maintainer}: Kevin Bosirany Orlando \email{kevinbosirany@gmail.com} (\href{https://orcid.org/0009-0009-2784-3108}{ORCID})

}
\keyword{internal}
